moved data structures to parser/datastructures.py
